refactor: rename apply earnings to apply profit fee
