+++ src/Core/BankingApp.Messaging.Contracts/IntegrationEvents.cs
@@ -9,6 +9,6 @@ public record AccountBalanceChangedIntegrationEvent(Guid HolderId, decimal Balan
 
 public record AccountOverdraftSettledIntegrationEvent(Guid HolderId, decimal OverdraftFee);
 
-public record AccountProfitFeeSettledIntegrationEvent(Guid HolderId, decimal OverdraftFee);
+public record AccountProfitFeeSettledIntegrationEvent(Guid HolderId, decimal ProfitFee);
 
 public record AccountEarningsIntegrationEvent(Guid HolderId, decimal Earnings);
