--- src/Transactions/BankingApp.Transactions.API/Features/ApplyEarnings/ApplyEarningsCommand.cs
@@ -1,7 +1,0 @@
-﻿using BankingApp.Domain.Core;
-using MediatR;
-
-// ReSharper disable ClassNeverInstantiated.Global
-namespace BankingApp.Transactions.API.Features.ApplyEarnings;
-
-public record ApplyEarningsCommand(Guid HolderId, decimal Earnings) : IRequest, ICommand;
