+++ src/Transactions/BankingApp.Transactions.API/Features/ApplyProfitFee/ApplyProfitFeeCommand.cs
@@ -1,0 +1,7 @@
+﻿using BankingApp.Domain.Core;
+using MediatR;
+
+// ReSharper disable ClassNeverInstantiated.Global
+namespace BankingApp.Transactions.API.Features.ApplyProfitFee;
+
+public record ApplyProfitFeeCommand(Guid HolderId, decimal Earnings) : IRequest, ICommand;
