+++ src/Transactions/BankingApp.Transactions.API/Features/ApplyProfitFee/ApplyProfitFeeCommandValidator.cs
@@ -1,0 +1,12 @@
+﻿using FluentValidation;
+
+namespace BankingApp.Transactions.API.Features.ApplyProfitFee;
+
+public class ApplyProfitFeeCommandValidator : AbstractValidator<ApplyProfitFeeCommand>
+{
+    public ApplyProfitFeeCommandValidator()
+    {
+        RuleFor(command => command.HolderId)
+            .Must(holderId => holderId != Guid.Empty);
+    }
+}
